指摘事項対応 - モデルを単数系に
